feat(BAC-18): added endpoint to return chat history
