--- packages/api/.env.example
@@ -1,46 +1,0 @@
-NODE_ENV=
-PORT=
-HOST=
-
-ME_CONFIG_MONGODB_ADMINUSERNAME=
-ME_CONFIG_MONGODB_ADMINPASSWORD=
-ME_CONFIG_MONGODB_URL=
-
-
-DATABASE_URI_DEVELOP=
-DATABASE_URI_PROD=
-DATABASE_URI_TEST=
-
-CLIENT_ID=
-CLIENT_SECRET=
-SERVER_ROOT_URI=
-
-LINKEDIN_SECRET_ID=
-LINKEDIN_CLIENT_ID=
-CALLBACK_URL=
-
-OPENAI_API_KEY=
-SESSION_SECRET=
-
-JWT_SECRET=
-JWT_OPTION=
-
-FB_CLIENT_ID=
-FB_CLIENT_SERECT=
-FB_CALLBACK_URL=
-
-MODEL_KEY=
-API_KEY=
-FILE_SIZE=104857
-
-SENDGRID_HOST=
-SENDGRID_USER=
-SENDGRID_PASSWORD=
-SENDGRID_PORT=
-
-
-ACCESSKEYID=
-S3SECRETEKEY=
-GRITTYBUCKETNAME=
-
-ASSEMBLYAI_API_KEY =
