+++ packages/api/package.json
@@ -12,7 +12,7 @@
     "author": "Team Carpenter Grit",
     "license": "ISC",
     "dependencies": {
-        "@aws-sdk/client-s3": "^3.216.0", 
+        "@aws-sdk/client-s3": "^3.216.0",
         "@banana-dev/banana-dev": "^4.0.0",
         "@sendgrid/mail": "^7.7.0",
         "aws-sdk": "^2.1260.0",
@@ -25,8 +25,8 @@
         "express-async-errors": "^3.1.1",
         "express-session": "^1.17.3",
         "express-validator": "^6.14.2",
-        "form-data": "^4.0.0", 
-        "file-type": "^18.0.0", 
+        "file-type": "^18.0.0",
+        "form-data": "^4.0.0",
         "formidable": "^2.0.1",
         "google-auth-library": "^8.7.0",
         "joi": "^17.7.0",
@@ -34,14 +34,14 @@
         "mongodb": "^4.12.0",
         "mongoose": "^6.7.2",
         "multer": "^1.4.5-lts.1",
-        "multer-s3": "^3.0.1", 
+        "multer-s3": "^3.0.1",
         "node-fetch": "^2.6.7",
         "openai": "^3.1.0",
         "passport": "^0.6.0",
         "passport-apple": "^2.0.1",
         "passport-facebook": "^3.0.0",
         "passport-linkedin-oauth2": "^2.0.0",
-        "path": "^0.12.7", 
+        "path": "^0.12.7",
         "sass": "^1.56.1",
         "supertest": "^6.3.1",
         "swagger-ui-express": "^4.6.0",
