+++ packages/api/routes/chatHistoryRoute.js
@@ -1,0 +1,7 @@
+const express = require('express')
+const router = express.Router()
+const historyController = require('../controller/chatHistoryController')
+
+router.get('/chathistory/:userId', historyController )
+
+module.exports = router
